refactor(component-library): publish MUI augmentations and fix the typecheck script
The earlier commit moved the design-system Button/IconButton/Breadcrumbs
augmentations inline into themes/code.org/index.ts. That worked, but two
follow-on findings push for a cleaner pass:

1. The Typography variants (body3, body4, overline1-3, figcaption,
   strong, em, label1-4) are not apps-specific. They are wired into
   themes/code.org/index.ts's createTheme() and used by component-library
   itself (Accordion, ActionBlock, FullWidthActionBlock, Slider). They
   belong here, not duplicated in apps/src/types/mui.d.ts.
2. component-library's `yarn typecheck` was a silent no-op:
   `tsc --noEmit` against a solution-style root tsconfig.json with
   `files: []` and references to tsconfig.app.json/tsconfig.node.json
   compiles nothing without `-b`. CI never caught the broken state.

Changes:

- Move all MUI augmentations (Button, IconButton, Breadcrumbs, barrel
  ButtonOwnProps, Typography variants + overrides) into a dedicated
  `themes/code.org/muiAugmentation.ts`. Surface via
  `export type * from './muiAugmentation'` in the theme entry — the
  re-export is preserved through vite-plugin-dts emission, while plain
  side-effect imports and triple-slash directives are stripped. A
  single sentinel type (`DesignSystemMuiAugmentations = void`) exists
  so the re-export has a named target (otherwise import-x/export
  flags it).
- Delete apps/src/types/mui.d.ts entirely. apps now picks up every
  augmentation transitively through @code-dot-org/component-library/themes.
- Wire the typecheck script to actually run: add `composite: true` to
  tsconfig.app.json and tsconfig.node.json, change scripts to
  `"build": "vite build"` and `"typecheck": "tsc -b --noEmit"`.
  Matches apps/studio's working solution-style precedent and fits the
  turbo task graph (lint dependsOn: typecheck) cleanly.
- Fix two pre-existing typecheck errors that the broken script was
  hiding:
  - Slider.test.tsx: drop dead `icon: {iconName: ...}` props from
    leftButtonProps/rightButtonProps. Stale from cdfc5947c24, when
    the prop type changed from DSCO ButtonProps to MuiIconButtonProps.
    The test asserts click handlers via aria-label; icon contents
    were never observed.
  - iconButton.tsx: annotate ICON_BUTTON_OVERRIDES with
    `Components<Theme>['MuiIconButton']` so CSS literal types like
    `position: 'relative'` keep their narrow types under contextual
    typing. Other style-override consts in this folder already do.

Verified: cd frontend/packages/component-library && yarn typecheck —
clean. cd frontend && yarn release:dryrun — 31/31. cd apps && yarn
typecheck — clean (Typography augmentations now flow in via
component-library's published types).
