+++ frontend/packages/component-library/tsconfig.app.json
@@ -2,6 +2,7 @@
   "extends": "@code-dot-org/lint-config/typescript/tsconfig.vite.app.json",
   "compilerOptions": {
     "baseUrl": ".",
+    "composite": true,
     "verbatimModuleSyntax": false
   },
   "include": ["src", "types"]
