+++ frontend/packages/component-library/tsconfig.node.json
@@ -1,7 +1,8 @@
 {
   "extends": "@code-dot-org/lint-config/typescript/tsconfig.vite.node.json",
   "compilerOptions": {
-    "baseUrl": "."
+    "baseUrl": ".",
+    "composite": true
   },
   "include": ["vite.config.ts"]
 }
